allow for externally managed token files It may be preferrable to have an external mechanism manage the OAuth2 tokens, rather than having the sasl-xoauth2 plug-in do this for you.
To allow for this scenario, a new config parameter "external_token_manager"
was introduced. When set to true, sasl-xauth2 will use the "password" field
content as the name of a file that contains the current OAuth2 token for
user. No other configuration is needed for this plug-in, but an external
mechanism must be in place to keep the content of that token file current
at all times.