Restore sasl-xoauth2-tool.
And add `test-config` and `test-token-refresh` subcommands. This is more in
line with @dkg's original vision for this tool.

The source script itself is now a template which is expanded at build time so
that references to the config file and the testing tool are set based on the
install prefix, etc.

The testing tool moves from `$BINDIR/sasl-xoauth2-test-config` to
`$LIBDIR/sasl-xoauth2/test-config` to avoid cluttering `$PATH`.