WIP: sample manpage for sasl-xoauth2.conf(5)
This bit of markdown can be converted to an NROFF-formatted manpage with the following command

    pandoc --standalone --to man  docs/sasl-xoauth2.conf.5.md -o sasl-xoauth2.conf.5

I'm not confident enough in CMake to know how to integrate pandoc into
the standard build process, but i can integrate it pretty easily into
the debian packaging if it's shipped this way.

Obviously, the description of each option needs to be fleshed out.

I think it'd be nicer in the long term to generate this manpage from
the code that parses the config file itself, to be sure that it stays
in sync, but i don't have a straightforward proposal for how to do
that.