don't require client-secret for outlook token requests
Not all outlook token requests require secrets be provided, and in
fact in these cases the requests will fail with a 401 error if a
secret is provided.  This patch fixes this issue to not require the
client-secret argument for outlook.

A prompt will now only be provided if the argument to the
--client-secret option is ":".

closes #64